Restricts Migrator to Database types only
Drop support for `DB::Connection` for `db` in Migrator in order to
simplify the types used by it.

This helps to reduce the confusion on what to pass to Migrator but also
helps with type resolution as `DB::Database | DB::Connection` was
presenting some issues when dealing with DB::Database and DB::Connection
method overload.